feat(gke): Break out gke specific code into own package
