+++ k8s/kube_config_test.go
@@ -1,4 +1,4 @@
-package handlers
+package k8s
 
 import (
 	"encoding/base64"
@@ -6,74 +6,45 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/arschles/assert"
-	"github.com/deis/k8s-claimer/clusters"
-	"github.com/deis/k8s-claimer/gke"
-	"github.com/deis/k8s-claimer/k8s"
-	"github.com/deis/k8s-claimer/leases"
-	"github.com/deis/k8s-claimer/testutil"
 	container "google.golang.org/api/container/v1"
-	"gopkg.in/yaml.v2"
+
+	yaml "gopkg.in/yaml.v2"
+
+	"github.com/arschles/assert"
 )
 
-const (
-	projID = "test project"
-	zone   = "test zone"
-)
-
-func TestFindUnusedGKEClusterByName(t *testing.T) {
-	leaseableClusters := testutil.GetClusters()
-
-	// test when all clusters are leased
-	fakeLister := &gke.FakeClusterLister{
-		Resp: &container.ListClustersResponse{Clusters: leaseableClusters},
-		Err:  nil,
+func TestCreateKubeClientFromConfig(t *testing.T) {
+	conf := &KubeConfig{
+		Kind:        "testKind",
+		APIVersion:  "v1",
+		Preferences: Preferences{Colors: true},
+		Clusters: []NamedCluster{
+			NamedCluster{
+				Name: "test1",
+				Cluster: Cluster{
+					Server:                   "test.server.com",
+					APIVersion:               "v1",
+					InsecureSkipTLSVerify:    false,
+					CertificateAuthorityData: ca,
+				},
+			},
+		},
+		AuthInfos: []NamedAuthInfo{
+			NamedAuthInfo{
+				Name: "test1",
+				AuthInfo: AuthInfo{
+					ClientCertificateData: pubKey,
+					ClientKeyData:         privKey,
+					Impersonate:           "impersonate1",
+					Username:              "testUser",
+					Password:              "testPass",
+				},
+			},
+		},
 	}
-
-	clusterMap, err := clusters.ParseMapFromGKE(fakeLister, projID, zone)
+	cl, err := CreateKubeClientFromConfig(conf)
 	assert.NoErr(t, err)
-	leaseMap, err := leases.ParseMapFromAnnotations(nil)
-
-	unusedCluster, err := findUnusedGKECluster(clusterMap, leaseMap, "getClusterByName", "")
-	assert.NoErr(t, err)
-	assert.Equal(t, unusedCluster.Name, "getClusterByName", "free cluster name")
-}
-
-func TestFindUnusedGKEClusterByVersion(t *testing.T) {
-	leaseableClusters := testutil.GetClusters()
-
-	// test when all clusters are leased
-	fakeLister := &gke.FakeClusterLister{
-		Resp: &container.ListClustersResponse{Clusters: leaseableClusters},
-		Err:  nil,
-	}
-
-	clusterMap, err := clusters.ParseMapFromGKE(fakeLister, projID, zone)
-	assert.NoErr(t, err)
-	leaseMap, err := leases.ParseMapFromAnnotations(nil)
-
-	unusedCluster, err := findUnusedGKECluster(clusterMap, leaseMap, "", "1.1.1")
-	assert.NoErr(t, err)
-	assert.Equal(t, unusedCluster.Name, "getClusterByVersion", "free cluster name")
-	assert.Equal(t, unusedCluster.CurrentNodeVersion, "1.1.1", "free cluster version")
-}
-
-func TestFindRandomUnusedGKECluster(t *testing.T) {
-	leaseableClusters := testutil.GetClusters()
-
-	// test when all clusters are leased
-	fakeLister := &gke.FakeClusterLister{
-		Resp: &container.ListClustersResponse{Clusters: leaseableClusters},
-		Err:  nil,
-	}
-
-	clusterMap, err := clusters.ParseMapFromGKE(fakeLister, projID, zone)
-	assert.NoErr(t, err)
-	leaseMap, err := leases.ParseMapFromAnnotations(nil)
-
-	unusedCluster, err := findUnusedGKECluster(clusterMap, leaseMap, "", "")
-	assert.NoErr(t, err)
-	assert.NotNil(t, unusedCluster, "free cluster name")
+	assert.NotNil(t, cl, "kube client")
 }
 
 func TestCreateKubeConfigFromCluster(t *testing.T) {
@@ -88,7 +59,7 @@ func TestCreateKubeConfigFromCluster(t *testing.T) {
 			Username:             "test username",
 		},
 	}
-	k8sConfig, err := createKubeConfigFromCluster(cluster)
+	k8sConfig, err := CreateKubeConfigFromCluster(cluster)
 	assert.NoErr(t, err)
 	assert.Equal(t, k8sConfig.APIVersion, "v1", "api version")
 	assert.Equal(t, k8sConfig.CurrentContext, strings.ToLower(cluster.Name), "context")
@@ -108,12 +79,6 @@ func TestCreateKubeConfigFromCluster(t *testing.T) {
 	assert.Equal(t, contextConfig.Context.AuthInfo, cluster.Name, "auth info name")
 }
 
-type contextClusterAndAuthInfo struct {
-	contextName  string
-	clusterName  string
-	authInfoName string
-}
-
 func TestMarshalAndEncodeKubeConfig(t *testing.T) {
 	const namespace = "myns"
 	const locationOfOrigin = "myloc"
@@ -122,63 +87,63 @@ func TestMarshalAndEncodeKubeConfig(t *testing.T) {
 	clusterNames := []string{"cluster1", "cluster2", "cluster3"}
 	authInfoNames := []string{"authInfo1", "authInfo2", "authInfo3"}
 
-	var contexts []k8s.NamedContext
+	var contexts []NamedContext
 	for i, contextName := range contextNames {
-		context := k8s.Context{
+		context := Context{
 			Cluster:   clusterNames[i],
 			AuthInfo:  authInfoNames[i],
 			Namespace: namespace,
 		}
-		namedContext := k8s.NamedContext{
+		namedContext := NamedContext{
 			Name:    contextName,
 			Context: context,
 		}
 		contexts = append(contexts, namedContext)
 	}
 
-	var clusters []k8s.NamedCluster
+	var clusters []NamedCluster
 	for _, clusterName := range clusterNames {
-		cluster := k8s.Cluster{
+		cluster := Cluster{
 			Server:                   clusterName + "/server",
 			APIVersion:               kubeconfigAPIVersion,
 			InsecureSkipTLSVerify:    false,
 			CertificateAuthorityData: clusterName + "_cert_authority",
 		}
-		namedCluster := k8s.NamedCluster{
+		namedCluster := NamedCluster{
 			Name:    clusterName,
 			Cluster: cluster,
 		}
 		clusters = append(clusters, namedCluster)
 	}
 
-	var authInfos []k8s.NamedAuthInfo
+	var authInfos []NamedAuthInfo
 	for _, authInfoName := range authInfoNames {
-		authInfo := k8s.AuthInfo{
+		authInfo := AuthInfo{
 			ClientCertificateData: authInfoName + "_cert_data",
 			ClientKeyData:         authInfoName + "_key_data",
 			Username:              authInfoName + "_username",
 			Password:              authInfoName + "_password",
 			Token:                 authInfoName + "_bearer_token",
 		}
-		namedAuthInfo := k8s.NamedAuthInfo{
+		namedAuthInfo := NamedAuthInfo{
 			Name:     authInfoName,
 			AuthInfo: authInfo,
 		}
 		authInfos = append(authInfos, namedAuthInfo)
 	}
 
-	cfg := &k8s.KubeConfig{
+	cfg := &KubeConfig{
 		CurrentContext: "ctx1",
 		Clusters:       clusters,
 		Contexts:       contexts,
 		AuthInfos:      authInfos,
 	}
 
-	yamlString, err := marshalAndEncodeKubeConfig(cfg)
+	yamlString, err := MarshalAndEncodeKubeConfig(cfg)
 	assert.NoErr(t, err)
 	decodedBytes, err := base64.StdEncoding.DecodeString(yamlString)
 	assert.NoErr(t, err)
-	decodedCfg := new(k8s.KubeConfig)
+	decodedCfg := new(KubeConfig)
 	assert.NoErr(t, yaml.Unmarshal(decodedBytes, decodedCfg))
 	assert.Equal(t, decodedCfg.APIVersion, cfg.APIVersion, "API version")
 	assert.Equal(t, decodedCfg.AuthInfos[0].AuthInfo.ClientCertificateData, cfg.AuthInfos[0].AuthInfo.ClientCertificateData,
