+++ k8s/namespace_lister.go
@@ -1,8 +1,6 @@
 package k8s
 
-import (
-	"k8s.io/client-go/pkg/api/v1"
-)
+import "k8s.io/client-go/pkg/api/v1"
 
 // NamespaceLister is a (k8s.io/kubernetes/pkg/client/unversioned).NamespaceInterface compatible
 // interface designed only for listing namespaces. It should be used as a parameter to functions
@@ -20,4 +18,9 @@ type FakeNamespaceLister struct {
 // List is the NamespaceLister interface implementation. It just returns f.NsList, f.Err
 func (f FakeNamespaceLister) List(opts v1.ListOptions) (*v1.NamespaceList, error) {
 	return f.NsList, f.Err
+}
+
+// NewFakeNamespaceLister returns a Fake NamespaceLister struct
+func NewFakeNamespaceLister(nsList *v1.NamespaceList, err error) *FakeNamespaceLister {
+	return &FakeNamespaceLister{NsList: nsList, Err: err}
 }
