+++ k8s/namespace_lister_deleter.go
@@ -1,5 +1,7 @@
 package k8s
 
+import "k8s.io/client-go/pkg/api/v1"
+
 // NamespaceListerDeleter contains both the listing and deleting capabilities of a
 // (k8s.io/kubernetes/pkg/client/unversioned).NamespaceInterface
 type NamespaceListerDeleter interface {
@@ -12,4 +14,19 @@ type NamespaceListerDeleter interface {
 type FakeNamespaceListerDeleter struct {
 	*FakeNamespaceLister
 	*FakeNamespaceDeleter
+}
+
+// GetNSFunc returns a function for NamespaceListerDeleter
+func GetNSFunc(nsListerDeleter NamespaceListerDeleter, err error) func(*KubeConfig) (NamespaceListerDeleter, error) {
+	return func(*KubeConfig) (NamespaceListerDeleter, error) {
+		return nsListerDeleter, err
+	}
+}
+
+// NewFakeNamespaceListerDeleter returns a new FakeNamespaceListerDeleter
+func NewFakeNamespaceListerDeleter(listNs *v1.NamespaceList, listErr error, deleteErr error) *FakeNamespaceListerDeleter {
+	return &FakeNamespaceListerDeleter{
+		FakeNamespaceLister:  NewFakeNamespaceLister(listNs, listErr),
+		FakeNamespaceDeleter: NewFakeNamespaceDeleter(deleteErr),
+	}
 }
