+++ providers/gke/cluster_lister.go
