+++ providers/gke/fake_cluster_lister.go
@@ -1,8 +1,6 @@
 package gke
 
-import (
-	container "google.golang.org/api/container/v1"
-)
+import container "google.golang.org/api/container/v1"
 
 // FakeClusterLister is a ClusterLister implementation for use in unit tests
 type FakeClusterLister struct {
@@ -13,4 +11,12 @@ type FakeClusterLister struct {
 // List is the ClusterLister interface implementation. It just returns f.Resp, f.Err
 func (f FakeClusterLister) List(projectID, zone string) (*container.ListClustersResponse, error) {
 	return f.Resp, f.Err
+}
+
+// NewFakeClusterLister returns a new FakeClusterLister
+func NewFakeClusterLister(resp *container.ListClustersResponse, err error) *FakeClusterLister {
+	return &FakeClusterLister{
+		Resp: resp,
+		Err:  err,
+	}
 }
