+++ providers/gke/gke_cluster_lister.go
